account for finalizers on Delete

+++ cmd/aks-node-viewer/main.go
@@ -165,7 +165,7 @@ func startMonitor(ctx context.Context, settings *monitorSettings) {
 			},
 			UpdateFunc: func(oldObj, newObj interface{}) {
 				n := newObj.(*v1.Node)
-				if !n.DeletionTimestamp.IsZero() {
+				if !n.DeletionTimestamp.IsZero() && len(n.Finalizers) == 0 {
 					cluster.DeleteNode(n.Name)
 				} else {
 					node, ok := cluster.GetNode(n.Name)
